Added ramfs and change stdio.h
I decided to finally add an actual file system. I was working on one for
a while that actually used the disk and everything. But I got bored.

+++ 32-bit/mods/std/include/graphics/image_struct.h
@@ -3,7 +3,8 @@
 
 typedef struct IntegerRange {
     uint8_t continueWith;
-    uint8_t continueFor;    
+    uint16_t continueFor; // When the image is used change this back to `uint8_t` below.
+    // uint8_t continueFor;    
 } IntegerRange;
 
 #endif
